requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.7
django-allauth==65.13.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
twilio==8.11.1
ua-parser==0.18.0
user-agents==2.2.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openapi-codec==1.3.2
pillow==12.0.0
PyYAML==6.0.3
requests==2.32.3
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
